# БР1.2/Беломытцев Андрей/labs/lab2/src/controllers/Review.ts
import { AppDataSource } from "../AppDataSource"
import { Review } from '../models/Review'
import { Controller, Get, Post, Delete, Route, Tags, Body, Path, Security, Request } from 'tsoa'
import { ReviewCreateDto, ReviewDto } from '../dto/Review';
import { User } from '../models/User'

const repository = AppDataSource.getRepository(Review)

@Tags('Review')
@Route('review')
export class ReviewController extends Controller {
  @Get()
  public async get(): Promise<ReviewDto[]> {
    var reviews = await repository.find({ relations: ['channel', 'user'] })
    return reviews.map((review) => { return { ...review, channelId: review.channel.id, userId: review.user.id } })
  }

  @Get('{id}')
  public async getOne(@Path() id: number): Promise<ReviewDto | null> {
    var review = await repository.findOne({ where: { id }, relations: ['channel', 'user'] })
    if (!review) return null
    return { ...review, channelId: review.channel.id, userId: review.user.id }
  }

  @Post()
  @Security('jwt')
  public async create(@Body() body: ReviewCreateDto, @Request() req: any): Promise<ReviewDto> {
    var review = await repository.save({...body, user: (await AppDataSource.getRepository(User).findOneBy({ username: req.user.username }))!})
    return { ...review, userId: review.user.id }
  }
  
  @Delete('{id}')
  @Security('jwt', ['admin'])
  public async remove(@Path() id: number) {
    const r = await repository.delete(id)
    if (r.affected === 0) {
      this.setStatus(404)
      throw new Error('Not found')
    }
    return r
  }
}