# БР1.2/Беломытцев Андрей/labs/lab2/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"DeleteResult": {
				"description": "Result object returned by DeleteQueryBuilder execution.",
				"properties": {
					"raw": {
						"description": "Raw SQL result returned by executed query."
					},
					"affected": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "Number of affected rows/documents\nNot all drivers support this"
					}
				},
				"required": [
					"raw"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ReviewDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"channelId": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"timeCreate": {
						"type": "string",
						"format": "date-time"
					},
					"text": {
						"type": "string"
					},
					"rate": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"channelId",
					"userId",
					"timeCreate",
					"text",
					"rate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewCreateDto": {
				"properties": {
					"channelId": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"rate": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"channelId",
					"text",
					"rate"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "lab1",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/review": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ReviewDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewDto"
								}
							}
						}
					}
				},
				"tags": [
					"Review"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewCreateDto"
							}
						}
					}
				}
			}
		},
		"/review/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ReviewDto"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteResult"
								}
							}
						}
					}
				},
				"tags": [
					"Review"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}